Extract types to individual files.
