+++ unity/unity-embed-host/AssemblyLoadContextUnloadException.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace Unity.CoreCLRHelpers;
+
+internal class AssemblyLoadContextUnloadException : Exception
+{
+}
